refactor(api): extract auth.Identity to auth/types leaf package (tv8.30)
Move the auth.Identity value type into a new leaf sub-package
encore.app/auth/types so consumers that only need the type (rbac
builder, future B-1..D-3 / E-* shared/* tests) can import it without
triggering auth's package-level sqldb.NewDatabase("unblock", ...) init,
which panics outside the encore CLI's Docker-backed cluster bring-up.

Approach: option (a) per orchestrator DIRECTIVE on unblock-tv8.30 (user
instruction 2026-05-08). The parent auth package re-exports the type
via `type Identity = types.Identity`, so the SPEC §10.1 surface
declaration `func For[T any](identity auth.Identity, ...)` continues
to be honoured at every Encore-service consumer call site (org.go,
future workitems/deps/mcp/...). The two spellings alias to the same
Go type — values constructed as `auth.Identity{...}` are
interchangeable with `types.Identity` parameters.

Changes:
- Add apps/api/auth/types/types.go with the Identity struct (locked
  shape per SPEC §4.1) and an extensive package doc explaining the
  Encore-free constraint.
- apps/api/auth/auth.go: replace the inline Identity declaration with
  `type Identity = types.Identity` re-export.
- apps/api/shared/rbac/rbac.go: switch the import from encore.app/auth
  to encore.app/auth/types; type the ScopedQuery.identity field and
  For[T]'s parameter as types.Identity. Package doc updated to
  explain the spelling reconciliation.
- apps/api/shared/rbac/rbac_test.go: switch fakeIdentity to
  types.Identity. Fix the two pre-existing assertion bugs in
  TestWhere_InjectionDocumentation_NoRuntimeSanitiser that were
  masked while the import-time panic prevented any test from running:
  the renumberer rewrites user `$1` to `$2` (because the scope
  predicate consumes $1), so the assertions now pin the meta-character
  bytes that follow the placeholder rather than the placeholder index.
  TestWhere_InjectionDocumentation_HostileClauseEmitsVerbatim already
  used a substring assertion and remains correct.

SPEC §3.1 invariant preserved: auth is still the sole migration-owner;
sqldb.NewDatabase("unblock", ...) is still called exactly once across
the workspace, in apps/api/auth/db.go. The new auth/types sub-package
imports zero encore.dev/* packages and declares no infrastructure.

Verification:
- `go test ./apps/api/shared/rbac/...` (no encore CLI, no Docker)
  passes — primary acceptance criterion. Was a panic before this
  patch ("encore apps must be run using the encore command").
- `go test ./apps/api/auth/types/...` passes (no test files; smoke
  compile only).
- `go test -race ./apps/api/shared/...` passes.
- `go fmt ./...` clean; `go vet ./...` clean; `go build ./...` clean.
- `encore check` builds the application graph and compiles the source
  tree successfully — Encore's parser accepts the auth/types
  sub-package layout (DRIFT-B fallback to shared/authtypes/ NOT
  needed). The Docker-cluster bring-up step still fails locally
  because Docker is not installed, which is exactly the situation
  this bead exists to address; CI runs encore check with Docker.

Constraints honoured per orchestrator DIRECTIVE:
1. SPEC §10.1 spelling auth.Identity preserved at every Encore-
   service consumer call site via the type alias. (rbac.go itself
   spells the type as types.Identity to break the import-time
   hazard chain — it is the implementation of the surface, not a
   service consumer; the alias makes the two spellings the same Go
   type so org.go's `auth.Identity{...}` literals interoperate
   seamlessly.)
2. Layout: apps/api/auth/types/ — confirmed working via encore
   check. Shared-package fallback (shared/authtypes/) not needed.
3. rbac.Bind() injection setter from unblock-tv8.4 DEVIATION stays.
4. SPEC §3.1 sole-migration-owner invariant preserved.
